+++ Changelog.md
@@ -12,6 +12,7 @@ Compiler Features:
 Bugfixes:
  * Optimizer: Fixed a bug in BlockDeDuplicator.
  * Type Checker: Disallow assignments to storage variables of type ``mapping``.
+ * Type Checker: Fix type deduction for exponentiation and shifts with literals as base.
 
 
 ### 0.6.8 (2020-05-14)